Massive changes to the frontpage
Frontpage is massively edited and soon finished. Script for data handling is beginning to match our purpose better. All of the files are copies from an earlier project that has been a little bit edited so far. Much of it will be gone and heavily edited by the end of this project.